backend/Penzi/requirements.txt
mysql-connector-python
flask_cors 
requests
flask_paginate 